Module Ppx_factory_lib.Default

val expr_from_core_type : loc:Ppxlib.Location.t -> Ppxlib.core_type -> (Ppxlib.expression, Loc_err.t) Stdlib.result

Return the default expression for the given core type, located at loc or a located error if a default expression can't be deduced. E.g. expr_from_core_type ~loc [%type int] is [%expr 0].

val expr_from_core_type_exn : loc:Ppxlib.Location.t -> Ppxlib.core_type -> Ppxlib.expression

Same as expr_from_core_type but raises instead of returning a result

val from_str_type_decl : (Ppxlib.structure, Ppxlib.rec_flag * Ppxlib.type_declaration list) Ppxlib.Deriving.Generator.t

Structure generator

val from_sig_type_decl : (Ppxlib.signature, Ppxlib.rec_flag * Ppxlib.type_declaration list) Ppxlib.Deriving.Generator.t

Signature generator

val _name_from_type_name : string -> string

Return the name of the default value derived from a type with the given name.