Module Util.Expr

val var : loc:Ppxlib.Location.t -> string -> Ppxlib.expression

Return the expression corresponding to the given variable name

val constructor : loc:Ppxlib.Location.t -> constructor_name:string -> Ppxlib.expression option -> Ppxlib.expression

Return the contructor expression with the given constructor name and argument expression