Module Ppx_factory_lib__Raise

Functions to raise ppx errors in ppx_factory

The loc argument should be the loc of the problematic node within the type declaration and not the loc argument of the generator to provide the user accurate information as to which part of the type declaration can't be handled.

val errorf : loc:Ppxlib.Location.t -> ('a, unit, string, 'b) Stdlib.format4 -> 'a

Raise an error with the formatted message prefixed by "ppx_factory: ".

module Default : sig ... end

module Factory : sig ... end