Module Ppx_factory_lib__Raise.Factory

val errorf : loc:Ppxlib.Location.t -> ('a, unit, string, 'b) Stdlib.format4 -> 'a

Raise an error with the formatted message prefixed by "ppx_factory.factory: "

val unhandled_type_kind : loc:Ppxlib.Location.t -> string -> 'a

Use when trying to derive a factory for an unhandled type kind. The message indicates factory can only be derived from record and variant types.