Module Ppx_factory_lib__Util

val affix_from_type_name : kind:[ `Prefix | `Suffix ] -> string -> string

Return the suffix to apply to a derived value name, based on the type name

val core_type_from_type_decl : loc:Ppxlib.Location.t -> Ppxlib.type_declaration -> Ppxlib.core_type

Return the core type describing the type declared in the given declaration. E.g. will return [%type: 'a t] for type 'a t = A of int | B of 'a.

val is_ocamldep : Ppxlib.Expansion_context.Deriver.t -> bool

Return whether the deriver is used in the context of ocamldep from the given expansion context

module Expr : sig ... end

module List_ : sig ... end

module Result_ : sig ... end